+++ CHANGES.md
@@ -18,6 +18,8 @@ E-Maj - Change log
     * Report a clear error message when using functions that read from or write
       to external files and the emaj installer role has not the appropriate
       rights.
+  * Use the dblink_connect() function when emaj is installed by a superuser,
+      avoiding to grant rights to dblink_connect_u() in most cases.
   * Improve the initial comment of the JSON output produced by
     emaj_export_groups_configuration().
   * Improve checks of the input JSON structure when importing a tables group